Instructions for uploading to PyPI
https://pypi.org/project/zaml/0.0.5/
